+++ extensions/robospice-google-http-client-parent/robospice-google-http-client/src/main/java/com/octo/android/robospice/JacksonGoogleHttpClientSpiceService.java
@@ -1,0 +1,22 @@
+package com.octo.android.robospice;
+
+import android.app.Application;
+
+import com.octo.android.robospice.persistence.CacheManager;
+import com.octo.android.robospice.persistence.exception.CacheCreationException;
+import com.octo.android.robospice.persistence.googlehttpclient.json.JacksonObjectPersisterFactory;
+
+/**
+ * A {@link GoogleHttpClientSpiceService} dedicated to json web services via
+ * Jackson. Provides caching.
+ * @author sni
+ */
+public class JacksonGoogleHttpClientSpiceService extends GoogleHttpClientSpiceService {
+    @Override
+    public CacheManager createCacheManager(Application application) throws CacheCreationException {
+        CacheManager cacheManager = new CacheManager();
+        cacheManager.addPersister(new JacksonObjectPersisterFactory(application));
+        return cacheManager;
+    }
+
+}