+++ robospice-core-parent/robospice/src/main/java/com/octo/android/robospice/SpiceManager.java
@@ -1231,5 +1231,7 @@ public class SpiceManager implements Runnable {
         public Exception getException() {
             return exception;
         }
+
     }
+
 }